changed code that relied on register_globals urlencode delete link for weird file names
git-svn-id: https://svn.radicaldesigns.org/amp/trunk@1231 75c0e4a9-b6e4-0310-a6ec-d70c6baf2916

+++ system/docdir.php
@@ -7,7 +7,7 @@ $listtitle ="Documents";
 $filename="docdir.php";
 
 //delete file 
-if ( isset($_GET['actdel']) ){
+if ( isset($_GET['actdel']) && $actdel=$_GET['actdel'] ){
 	$dir_name=AMP_LOCAL_PATH."/downloads/";
 	unlink($dir_name.$actdel);
 }
@@ -40,7 +40,7 @@ for($t=0;$t<$numberOfFiles;$t++){
 	
 	$fieldsarray[$t]['Document'] = '<a href="../downloads/'.$thisName.'" target="_blank">'.$thisName.'</a>';
 	$fieldsarray[$t]['Date'] = $thisTime;
-	$fieldsarray[$t]['Delete'] = '<a href="docdir.php?actdel='.$thisName.'">delete</a>';
+	$fieldsarray[$t]['Delete'] = '<a href="docdir.php?actdel='.urlencode($thisName).'">delete</a>';
 }
 closedir ($dir);
